+++ libs/sdk/skyux-stylelint/src/index.ts
@@ -1,4 +1,5 @@
+import noNgDeep from './rules/no-ng-deep.js';
 import noSkySelectors from './rules/no-sky-selectors.js';
 import noStaticColorValues from './rules/no-static-color-values.js';
 
-export default [noSkySelectors, noStaticColorValues];
+export default [noNgDeep, noSkySelectors, noStaticColorValues];